mcp/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
ollama
mcp